feat(strategies): added regex matches replacements for proxy response
