+++ .gitignore
@@ -4,3 +4,5 @@ vendor/
 .project
 .settings
 composer.lock
+tests/_output
+tests/_support/_generated
